server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
aiohttp==3.9.1
shapely==2.0.2
pyproj==3.6.1
python-dotenv==1.0.0